// client/src/Tools/data.model.ts
// Generated by https://quicktype.io

export interface JSONData {
    technologies: Technology[];
    courses: Courses[];
}

export interface Technology {
    _id:         string;
    name:        string;
    description: string;
    difficulty:  number;
    courses:     Course[];
}

export interface Courses {
    _id:  string;
    code: string;
    name: string;
}

export interface Course {
    code: string;
    name: string;
}

export interface ComponentProps {
    technologies:Technology[];
    courses: Courses[];
    fetchData:Function;
    setLoading:Function;
}
export interface LoadingProps {
    spinnerColor:string;
    bgColor:string
    enabled:boolean; 
}